Move PasswordStrengthTextField.tsx file and fix colors a little bit
