+++ src/components/Dialog/ChangePassword.tsx
@@ -9,8 +9,8 @@ import TextField from "@material-ui/core/TextField"
 import LockIcon from "@material-ui/icons/LockOutlined"
 import LockOpenIcon from "@material-ui/icons/LockOpenOutlined"
 import zxcvbn from "zxcvbn"
+import PasswordStrengthTextField from "../Form/PasswordStrengthTextField"
 import { Box, VerticalLayout } from "../Layout/Box"
-import PasswordStrengthTextField from "../Layout/PasswordStrengthTextField"
 import { Account, AccountsContextType } from "../../context/accounts"
 import { NotificationsContext } from "../../context/notifications"
 import { useIsMobile } from "../../hooks"
